requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pillow
qrcode
pyyaml
csscompressor
htmlmin2
dnspython
python-whois
requests
sqlparse
markdown
python-slugify
mangum
